requirements.txt
appdirs==1.4.3
colorama==0.3.7
Django==1.10.5
django-suit==0.2.24
django-templatetag-handlebars==1.3.1
djangorestframework==3.4.6
djangorestframework-expiring-authtoken==0.1.4
flake8==3.3.0
httpie==0.9.9
mccabe==0.6.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psutil==5.2.0
PyAutoGUI==0.9.33
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
PyMsgBox==1.0.4
pyparsing==2.2.0
PyScreeze==0.1.9
PyTweening==1.0.3
PyYAML==3.12
requests==2.13.0
six==1.10.0